+++ DesktopApplicationTemplate.Tests/CsvServiceTests.cs
@@ -21,7 +21,7 @@ namespace DesktopApplicationTemplate.Tests
         public void EnsureColumnsForService_AddsColumns()
         {
             var configPath = Path.Combine(Path.GetTempPath(), Guid.NewGuid().ToString()+".json");
-            var vm = new CsvViewerViewModel(configPath);
+            var vm = new CsvViewerViewModel(new StubFileDialogService(), configPath);
             var svc = new CsvService(vm);
 
             svc.EnsureColumnsForService("TestSvc");
@@ -37,7 +37,7 @@ namespace DesktopApplicationTemplate.Tests
         public void RemoveColumnsForService_RemovesColumns()
         {
             var configPath = Path.Combine(Path.GetTempPath(), Guid.NewGuid().ToString()+".json");
-            var vm = new CsvViewerViewModel(configPath);
+            var vm = new CsvViewerViewModel(new StubFileDialogService(), configPath);
             var svc = new CsvService(vm);
 
             svc.EnsureColumnsForService("Svc");
@@ -59,7 +59,7 @@ namespace DesktopApplicationTemplate.Tests
             try
             {
                 var configPath = Path.Combine(Path.GetTempPath(), Guid.NewGuid().ToString()+".json");
-                var vm = new CsvViewerViewModel(configPath);
+                var vm = new CsvViewerViewModel(new StubFileDialogService(), configPath);
                 vm.Configuration.FileNamePattern = Path.Combine(dir, "out_{index}.csv");
                 var svc = new CsvService(vm);
 
@@ -86,7 +86,7 @@ namespace DesktopApplicationTemplate.Tests
         public void RecordLog_WritesCsvRow()
         {
             var configPath = Path.Combine(Path.GetTempPath(), Guid.NewGuid().ToString()+".json");
-            var vm = new CsvViewerViewModel(configPath);
+            var vm = new CsvViewerViewModel(new StubFileDialogService(), configPath);
             string path = Path.Combine(Path.GetTempPath(), Guid.NewGuid().ToString()+".csv");
             vm.Configuration.FileNamePattern = path;
             var svc = new CsvService(vm);
@@ -109,7 +109,7 @@ namespace DesktopApplicationTemplate.Tests
             try
             {
                 var configPath = Path.Combine(Path.GetTempPath(), Guid.NewGuid().ToString()+".json");
-                var vm = new CsvViewerViewModel(configPath);
+                var vm = new CsvViewerViewModel(new StubFileDialogService(), configPath);
 
                 vm.Configuration.FileNamePattern = Path.Combine(tempDir, "output_{index}.csv");
                 var service = new CsvService(vm);
@@ -135,6 +135,36 @@ namespace DesktopApplicationTemplate.Tests
         [Fact]
         [TestCategory("CodexSafe")]
         [TestCategory("WindowsSafe")]
+        public void AppendRow_CreatesNestedDirectory()
+        {
+            var baseDir = Path.Combine(Path.GetTempPath(), Guid.NewGuid().ToString());
+            Directory.CreateDirectory(baseDir);
+            try
+            {
+                var configPath = Path.Combine(Path.GetTempPath(), Guid.NewGuid().ToString()+".json");
+                var vm = new CsvViewerViewModel(new StubFileDialogService(), configPath);
+                vm.Configuration.OutputDirectory = baseDir;
+                vm.Configuration.FileNamePattern = Path.Combine("TCP1.message", "out_{index}.csv");
+                var service = new CsvService(vm);
+
+                service.AppendRow(new[] { "x", "y" });
+
+                var expectedDir = Path.Combine(baseDir, "TCP1.message");
+                Assert.True(Directory.Exists(expectedDir));
+                var files = Directory.GetFiles(expectedDir, "out_0.csv");
+                Assert.Single(files);
+            }
+            finally
+            {
+                Directory.Delete(baseDir, true);
+            }
+
+            ConsoleTestLogger.LogPass();
+        }
+
+        [Fact]
+        [TestCategory("CodexSafe")]
+        [TestCategory("WindowsSafe")]
         public void AppendRow_WithoutIndexPlaceholder_UsesSingleFile()
         {
             var tempDir = Path.Combine(Path.GetTempPath(), Guid.NewGuid().ToString());
@@ -142,7 +172,7 @@ namespace DesktopApplicationTemplate.Tests
             try
             {
                 var configPath = Path.Combine(Path.GetTempPath(), Guid.NewGuid().ToString()+".json");
-                var vm = new CsvViewerViewModel(configPath);
+                var vm = new CsvViewerViewModel(new StubFileDialogService(), configPath);
 
                 var filePath = Path.Combine(tempDir, "output.csv");
                 vm.Configuration.FileNamePattern = filePath;
@@ -183,7 +213,7 @@ namespace DesktopApplicationTemplate.Tests
                     Directory.CreateDirectory(Path.GetDirectoryName(servicesPath)!);
                     var network = new Mock<INetworkConfigurationService>();
                     var networkVm = new NetworkConfigurationViewModel(network.Object);
-                    var mainVm = new MainViewModel(new CsvService(new CsvViewerViewModel(configPath)), networkVm, network.Object, null, servicesPath);
+                    var mainVm = new MainViewModel(new CsvService(new CsvViewerViewModel(new StubFileDialogService(), configPath)), networkVm, network.Object, null, servicesPath);
                     var view = new MainView(mainVm);
 
                     var svc = new ServiceViewModel { DisplayName = "CSV Creator - Test", ServiceType = "CSV Creator" };