+++ DesktopApplicationTemplate.Tests/MainViewModelFilterTests.cs
@@ -18,7 +18,7 @@ namespace DesktopApplicationTemplate.Tests
         public void NameFilter_FiltersServices()
         {
             var configPath = Path.Combine(Path.GetTempPath(), Guid.NewGuid().ToString() + ".json");
-            var csv = new CsvService(new CsvViewerViewModel(configPath));
+            var csv = new CsvService(new CsvViewerViewModel(new StubFileDialogService(), configPath));
             var network = new Mock<INetworkConfigurationService>();
             var networkVm = new NetworkConfigurationViewModel(network.Object);
             var vm = new MainViewModel(csv, networkVm, network.Object);